# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledgerful Daemon API",
    "description": "Machine-readable OpenAPI contract for the Ledgerful daemon `/api/*` endpoints. Generated from the Rust DTOs via utoipa.",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "`GET /api/events` — authenticated SSE stream of [`DaemonEvent`].",
        "description": "Wire format:\n- `event: daemon`\n- `data: <json DaemonEvent camelCase>`\n- Keep-alive comments every [`SSE_KEEPALIVE`]\n\nAuth: inherits `token_layer` (Bearer). Missing/invalid token → 403 before\nthe stream opens.",
        "operationId": "getEvents",
        "responses": {
          "200": {
            "description": "SSE stream of DaemonEvent frames (event name `daemon`)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonEvent"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid Bearer token"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DaemonEvent": {
        "type": "object",
        "description": "Narrow daemon status event pushed over `GET /api/events` (SSE).\n\nDrift counts + readiness only — deliberately excludes `model_reachable`\nprobes and `is_demo` (constant for a daemon session). See track 0085.",
        "required": [
          "pendingTransactions",
          "unauditedDrift",
          "indexReady",
          "graphReady"
        ],
        "properties": {
          "graphReady": {
            "type": "boolean"
          },
          "indexReady": {
            "type": "boolean"
          },
          "pendingTransactions": {
            "type": "integer",
            "minimum": 0
          },
          "unauditedDrift": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}